feat: put helpers together
